backend/requirements.txt
astroid==2.0.4
awscli==1.16.60
awsebcli==3.14.6
blessed==1.15.0
botocore==1.12.50
cached-property==1.5.1
cement==2.8.2
certifi==2018.10.15
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docker==3.5.1
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.9.0
idna==2.6
isort==4.3.4
jmespath==0.9.3
jsonschema==2.6.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.13
pathspec==0.5.5
psycopg2==2.7.6.1
pyasn1==0.4.4
PyJWT==1.7.1
pylint==2.1.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.1.0
urllib3==1.22
uWSGI==2.0.17.1
wcwidth==0.1.7
websocket-client==0.54.0
whitenoise==4.1.2
wrapt==1.10.11
